Extract counting mock connector
